ckps module refactoring for support 36-1 wheel. Also, performed changing of general types to: int8_t, uint8_t, int16_t, uint16_t, int32_t, uint32_t.
git-svn-id: http://svn2.assembla.com/svn/secu3app@85 c1dae7de-a428-49f4-85d8-50b65f41d01d

+++ crc16.c
@@ -19,10 +19,10 @@
 *  Возвращает
 *    unsigned short crc16 буфера
 */
-WORD crc16( BYTE *buf, WORD num )
+uint16_t crc16( uint8_t *buf, uint16_t num )
 {
-unsigned int i;
-unsigned short crc = 0xffff;
+uint16_t i;
+uint16_t crc = 0xffff;
 
   while ( num-- )
   { 
@@ -40,10 +40,10 @@ unsigned short crc = 0xffff;
 }
 
 //вариант для данных во flash
-WORD crc16f(BYTE __flash *buf, WORD num )
+uint16_t crc16f(uint8_t __flash *buf, uint16_t num )
 {
-unsigned int i;
-unsigned short crc = 0xffff;
+uint16_t i;
+uint16_t crc = 0xffff;
 
   while ( num-- )
   {
